Voting protocol: updates per 2020-08-03 meeting
This document has been present for some time, but was not updated to
reflect the charter once the charter was approved and accepted by the
LF.

The LF allows us to define our own processes, including how voting is
held, so long as those processes do not conflict with the charter. The
previous text conflicted with the charter. The proposed changes are in
line with the charter, and contain the [clarifications we voted on during the TSC meeting held 2020-08-03](https://github.com/laminas/technical-steering-committee/pull/39/files#diff-8b5332138fd8c5a5d1edf411fc2a0469R169-R176).

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>